requirements.txt
apache-airflow==2.7.0
mlflow==2.8.1
pandas==2.0.3
scikit-learn==1.3.0
kaggle==1.6.11
python-dotenv==1.0.0